// src/types/task.ts
export interface Task {
  _id: string;
  title: string;
  description: string;
  status: 'todo' | 'in-progress' | 'completed';
  priority: 'low' | 'medium' | 'high';
  dueDate: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskInput = Omit<Task, '_id' | 'createdAt' | 'updatedAt'>;